bench/time: cycle timing utilities
We use an asymmetric pair of CPUID/RDTSC and RDTSCP/CPUID as
recommended in
https://www.intel.com/content/dam/www/public/us/en/documents/white-papers/ia-32-ia-64-benchmark-code-execution-paper.pdf

We also detect context switches by filling %gs with an invalid segment
descriptor, which is reset to 0 by IRET
(https://lackingrhoticity.blogspot.com/2018/01/observing-interrupts-from-userland-on-x86.html).

It's important to detect context switches because we want to throw
away observations that were interrupted.  However, it's even more
important to report them, because they tend to introduce noise and
sometimes bias in the measurements.

TESTED=smoke test.