bench/internal/constructable-array: something like C arrays, but with destructors
We want to record values in the timing loop.  However, we want the
timing functions to be as small as possible; that means avoiding
default construction (which may not even be available) as well as
moving destruction to a shared function.

The ConstructableArray template achieves that, which makes it very
easy to misuse.

TESTED=smoke test.