bench/kolmogorov-smirnov-test: 3-way KS test
This test can not only reject the hypothesis that A is never slower
than B or vice versa, but can also determine when A is (probably)
*always* faster than B, or vice versa, modulo some leeway.

When A is always almost faster than B and B always almost faster than
A, we can even declare approximate equality!

Inversely, we can also determine that A is sometimes slower than B
and A sometimes faster than B; they two are then different, but not
comparable.

TESTED=new unit and integration tests.