bench/compare-functions: internal late-bound refactor
When we use this be benchmarking harness for regression testing, we'll
want to easily compare with functions at specific commits.  Copying
source, etc., is bound to be error-prone.  Instead, we'll generate
hermetic shared objects and extract one symbol from the old and new
.so files.

When we do that, we'll have to take a late-binding hit.  The harness
should make sure to absorb that hit outside the timed call by
including the full preparation / conversion and timing code in the
indirect callee.

The refactor in this commit is purely internal, and is minimal to help
bisection if we discover that late binding seems to introduce noise.

TESTED=existing tests pass.