bench: pass late-bound timing function bodies in a class with explicit vtables
This std::function substitute lets us pass around code that's compiled
with different runtimes, as long as the calls are ABI compatible.  In
constrast to std::function, this means we don't have to worry about
the host runtime destroying or copying functions created in the timed
libraries' runtime systems.

This change is purely internal, and is as small as possible to
facilitate bisection if using our own late-bound functor class
seems to introduce noise.

TESTED=existing tests pass.