bench/compare-libraries: compare hermetically linked shared objects
We generate a type-safe interface that uses only the C ABI to pass
code from hermetic shared objects back to the test harness, and make
that opaque code handle timing to avoid polluting timings with the
indirect call.

This lets us use the ExplicitFunction late-binding machinery to pass
code across shared object boundaries, and fully isolate shared objects
from the test hardness when possible.  Otherwise, we still get a lot
of isolation, but must share base libraries like libc and stdlibc++.

In a way, this change simplifies compare-functions.h, arguably by
scattering the complexity around.

The interface in the new compare-libraries.h is definitely easier to
misuse than compare-functions.h. The main reason we want all this is
to make it easier to compare the same function at different commits or
optimisation options. More ambitiously, this .so-based interface lets
us see what happens when we not only change compilers, but even when
we switch libc or C++ standard library!

Why not IPC between fully distinct executables? Mostly because they
give more opportunities to OS to consistently mess with our
timings. Multiple executables make it harder to share the same memory
for our inputs (which makes it possible that one process's heap is
accidentally on DIMMs that are closer or don't have noisy
neighbours). They also make it harder to make sure a pair of timed
values happen in comparable circumstances, e.g., on the same
code. Finally, we need batching to avoid polluting our measurements
with OS IPC code, which also makes it harder to randomise ordering
noise away.

We could nevertheless use the statistical testing code in a
multi-process environment. I'm just not sure we need that right now.

TESTED=new integration tests.