bench/bounded-mean-test: compare means modulo outliers, and outlier %
In general, we can't run statistical tests on means: nothing prevents
a distribution from having a marginal (1e-100) probabilty of returning
a really high (1e1000) value.  Means are just statistically unstable.

However, we can do something if we further add the condition that
we're only interested in values below a certain limit: everything
above that is an outlier and shouldn't count in the averages.  Once
the domain is bounded, we can compute confidence intervals for the
distribution averages of A and B, and compare these intervals.

Since we filter out outliers, we must look for differences in the rate
of outliers, but only if the one of the outlier rates is high enough
to warrant interest.

TESTED=new unit and integration tests.