bench/compare-functions: assert out when the functions to time go through indirect calls
This usually happens when passing a straight function pointer to
CompareFunctions.  We want to pass in a functor that itself calls the
function to time.  The functor's body will be inlined, and the call
will be direct.

And add an easy to use macro for this pattern in bench/wrap-function.h

TESTED=manually checked bad usage doesn't build, integration tests now use WRAP_FUNCTION.