bench/internal/dynamic-loading: opportunistically remap .text sections with huge pages
And make both remapping and dlmopen optional, with an Options struct.

Backing .text segments (executable machine code) with huge pages
introduces a bit of consistent bias and confuses perf, but greatly
improves consistency.  It's well worth the effort to control for
dl(m)open-ing order in test code.

In return for that reduced variance, disable dlmopen by default.
Hermetic copies of libc and stdlibc++ seem to help variance a bit, but
the downside is outright crashes and deadlocks when libc gets confused
by its clone.  Only enable this when we really need it.

TESTED=ran follow-up test code manually.