BUILD files: mark all binaries and libraries linkstatic
We only really care about cc_binary rules (either .so or executables),
which should be linked statically for performance and to be more
independent from the environment.  However, as we play visibility
tricks, intermediate libraries will fail to build their .so, even if
we only need their .a.  So, mark every target linkstatic.

TESTED=it builds.